Move `Cache` and `FeedImage` model into its own files
